Added filtering by dice when running card searching logic
As part of this update, I completely rewrote the way dice information is stored in the database. We are now using binary flags for cards, and card and die deck relationships now go through intermediary models (because previously there was no way to know the *count* of each card/die for the deck).