Reverted server-only filtering
Because life is just too damn short. I left in the Nanobar usage, though, so it still looks the same (and the experience on mobile devices where the filter takes forever should *act* like it's a server call, which will hopefully be a lot nice UX even if you still have to wait).